backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
blinker==1.9.0
    # via flask
click==8.1.8
    # via flask
flask==3.1.0
    # via
    #   -r requirements.in
    #   flask-cors
flask-cors==5.0.1
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mysql-connector-python==9.2.0
    # via -r requirements.in
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
